+++ packages/e2e-test-app/.eslintrc.js
@@ -2,8 +2,6 @@ module.exports = {
   extends: ['@rnw-scripts'],
   parserOptions: { tsconfigRootDir: __dirname },
   globals: {
-    $: 'readonly',
-    browser: 'readonly',
     expect: 'readonly',
     fail: 'readonly',
     AutomationClient: 'readonly',