+++ vnext/codegen/react/components/rnwcore/Props.h
@@ -10,10 +10,10 @@
 #pragma once
 
 #include <cinttypes>
-#include <react/components/view/ViewProps.h>
-#include <react/core/PropsParserContext.h>
-#include <react/graphics/Color.h>
-#include <react/imagemanager/primitives.h>
+#include <react/renderer/components/view/ViewProps.h>
+#include <react/renderer/core/PropsParserContext.h>
+#include <react/renderer/graphics/Color.h>
+#include <react/renderer/imagemanager/primitives.h>
 #include <vector>
 
 namespace facebook {